# perf(python): drop the pyarrow conversion path in `iter_rows`; we can now do fully native conversion ~2-3x faster